Tidying up and finalising on Windows.
