+++ package/README.md
@@ -1,0 +1,64 @@
+# Packaging
+
+Packaging datashuttle is slightly more complex than packaging a simple Python application,
+because we need to vendor the terminal to run datashuttle in, because it can render strangely
+depending on the platform (e.g. macOS terminal, older Windows terminals).
+
+Wezterm, a cross-platform terminal that renders datashuttle well. It is vendored as part of 
+this distribution.
+
+Packaging is a two-step process. First, we package datashuttle into an executable.
+This executable contains all datashuttle dependencies and can be run from any terminal.
+
+Next, a simple script ('terminal launcher') that runs the datashuttle executable in the vendored 
+Wezterm terminal is packaged using pyinstaller. This is the main entry point for the distribution.
+
+Finally, this executable is distributed via an installer (inno setup on Windows, TODO on macos (`.dmg`)
+This installs the executable in the proper place on the system, adds shortcuts etc.
+
+So the call chain is:
+1) The main executable is 'terminal launcher'
+2) This opens the vendored Wezterm, and uses it to start the second, datashuttle executable.
+
+## The `package` directory
+
+The path handling is a little fiddly here, and some scrips are shared  between platform. 
+Therefore, all materials are currently stored in this folder and not refactored into their own folder.
+
+The key files are:
+
+- **datashuttle.spec**: The pyinstaller command file that packages datashuttle, through `datashuttle_launcher.py` script.
+    This outputs a datashuttle executable that contains all datashuttle dependencies, and can be run through a terminal.
+- **datashuttle_launcher.py**: A wrapper script that `datashuttle.spec` packages, that launches datashuttle.
+- **license.txt**: The datashuttle license
+- **package_<macos_or_windows>.py**: The entry point script that coordinates the packaging of datashuttle on macOS / Windows.
+- **terminal_launcher_<macos_or_windows>.py**: The entry point script for the final datashuttle distribution. 
+    It opens Wezterm and runs the datashuttle executable in it.
+- **terminal_launcher_macos.spec**: The Pyinstaller command file that packages the `terminal_launcher` script into an executable.
+- **wezterm_config.lua**: The Wezterm config. Use this to change settings that affect how the vendored Wezterm appears.
+- **inno_compile_script.iss**:  The Inno script used to package the Windows distribution into an installer. The output
+    of this script is shipped.
+- **packaging_utils.py**: General utils used in the packaging scripts.
+
+## Windows
+
+When packaging on Windows, a number of intermediate folders will be created. Under normal circumstances
+they will be deleted, but knowing their contents can be useful for debugging:
+
+- **build**: Generated by pyinstaller, it contains intermediate files generating during the 
+packaging of `datashuttle` and `terminal_launcher`. Generally you do not need to inspect these files,
+they are pyinstaller stuff.
+- **dist**: Initially generated by pyinstaller, this is the complete distribution. During packaging,
+this file is used for the Pyinstaller dist of `datashuttle` and `terminal_launcher`, then we 
+copy a few auxillary files (e.g. license, Wezterm config)). Because we are building two Pyinstaller
+packages into one directory, some care had to be taken to avoid configs. The folders included are:
+1) `datashuttle`: The datashuttle distribution generated by Pyinstaller
+2) `terminal_launcher.exe` and `_internal` is the `terminal_launcher` distribution generated by Pyinstaller.
+3) `_vendored` is our vendor folder, that contains the vendored Wezterm temrinal.
+4) datashuttle license and icon
+- **Output**: This is the output of Inno Setup, containing the installer. This is the final version to be shipped.
+
+
+
+
+## macOS
