+++ package/make_inno_setup_script.py
@@ -1,0 +1,73 @@
+def make_inno_setup_script(version, base_path):
+    """{} are required in text so easier to use this method"""
+    file_body = (
+        r"""
+        ; Script generated by the Inno Setup Script Wizard.
+    ; SEE THE DOCUMENTATION FOR DETAILS ON CREATING INNO SETUP SCRIPT FILES!
+
+    #define MyAppName "Datashuttle"
+    #define MyAppVersion """
+        + '"'
+        + version
+        + """"
+    #define MyAppPublisher "Neuroinformatics Unit, Sainsbury Wellcome Centre, London, UK."
+    #define MyAppURL "www.datashuttle.neuroinformatics.dev"
+    #define MyAppExeName "terminal_launcher.exe"
+
+    [Setup]
+    ; NOTE: The value of AppId uniquely identifies this application. Do not use the same AppId value in installers for other applications.
+    ; (To generate a new GUID, click Tools | Generate GUID inside the IDE.)
+    AppId={{D063D197-9D21-426B-BF8C-D5612C74DF15}
+    AppName={#MyAppName}
+    AppVersion={#MyAppVersion}
+    ;AppVerName={#MyAppName} {#MyAppVersion}
+    AppPublisher={#MyAppPublisher}
+    AppPublisherURL={#MyAppURL}
+    AppSupportURL={#MyAppURL}
+    AppUpdatesURL={#MyAppURL}
+    DefaultDirName={autopf}\DataShuttle
+    DisableProgramGroupPage=yes
+    LicenseFile="""
+        + base_path
+        + """\dist\license.txt
+    ; Uncomment the following line to run in non administrative install mode (install for current user only.)
+    ;PrivilegesRequired=lowest
+    PrivilegesRequiredOverridesAllowed=dialog
+    OutputBaseFilename=datashuttle_"""
+        + version
+        + """
+    SetupIconFile="""
+        + base_path
+        + r"""\dist\NeuroBlueprint_icon.ico
+    UninstallDisplayIcon={app}\NeuroBlueprint_icon.ico
+    Compression=lzma
+    SolidCompression=yes
+    WizardStyle=modern
+
+    [Languages]
+    Name: "english"; MessagesFile: "compiler:Default.isl"
+
+    [Tasks]
+    Name: "desktopicon"; Description: "{cm:CreateDesktopIcon}"; GroupDescription: "{cm:AdditionalIcons}"; Flags: unchecked
+
+    [Files]
+    Source: """
+        + '"'
+        + base_path
+        + r"""\dist\terminal_launcher.exe"; DestDir: "{app}"; Flags: ignoreversion
+    Source: """
+        + '"'
+        + base_path
+        + """\dist\*"; DestDir: "{app}"; Flags: ignoreversion recursesubdirs createallsubdirs
+    ; NOTE: Don't use "Flags: ignoreversion" on any shared system files
+
+    [Icons]
+    Name: "{autoprograms}\{#MyAppName}"; Filename: "{app}\{#MyAppExeName}"
+    Name: "{autodesktop}\{#MyAppName}"; Filename: "{app}\{#MyAppExeName}"; Tasks: desktopicon
+
+    [Run]
+    Filename: "{app}\{#MyAppExeName}"; Description: "{cm:LaunchProgram,{#StringChange(MyAppName, '&', '&&')}}"; Flags: nowait postinstall skipifsilent
+    """
+    )
+
+    return file_body
